requirements.txt
requests
flask
wheel
uwsgi
query-string
psycopg2-binary
SQLAlchemy
Flask-Mail
gunicorn